changes to demo tokens

+++ programs/demo-tokens/src/state/protocol.rs
@@ -3,5 +3,5 @@ use anchor_lang::prelude::*;
 #[account]
 pub struct Protocol {
     pub bump: u8,
-    pub num_tickers: u8,
+    pub num_tickers: u64,
 }
